Fix Firefox bug in stats modal bottom padding
Firefox is collapsing paddings and margins, and for that reason it
looks different in other browsers.

Nothing seemed to work, not even adjusting the container space using
the ::after pseudo-element, adding a 'div', an 'hr', or using only
margin-top (which shouldn't be collapsed).

For that reason I decided to use '@supports' with a Firefox specific
property to target only this browser and make the necessary
adjustments.

More options and info about this technique can be see in:
http://browserhacks.com/